\name{data.demonfx}
\alias{demonfx}
\title{Demon FX Data Set}
\usage{data(demonfx)}
\description{
  This data set consists of daily currency pair prices from 2010 through
  2014. Each currency pair has a close, high, and low price.
}
\format{
  This data frame contains 1,301 rows as time-periods (with row names) and
  39 columns of currency pair prices. The following data dictionary describes
  each time-series or column.
  \describe{
    \item{\code{EURUSD.Close}}{This is the currency pair closing price.}
    \item{\code{EURUSD.High}}{This is the currency pair high price.}
    \item{\code{EURUSD.Low}}{This is the currency pair low price.}
    \item{\code{USDJPY.Close}}{This is the currency pair closing price.}
    \item{\code{USDJPY.High}}{This is the currency pair high price.}
    \item{\code{USDJPY.Low}}{This is the currency pair low price.}
    \item{\code{USDCHF.Close}}{This is the currency pair closing price.}
    \item{\code{USDCHF.High}}{This is the currency pair high price.}
    \item{\code{USDCHF.Low}}{This is the currency pair low price.}
    \item{\code{GBPUSD.Close}}{This is the currency pair closing price.}
    \item{\code{GBPUSD.High}}{This is the currency pair high price.}
    \item{\code{GBPUSD.Low}}{This is the currency pair low price.}
    \item{\code{USDCAD.Close}}{This is the currency pair closing price.}
    \item{\code{USDCAD.High}}{This is the currency pair high price.}
    \item{\code{USDCAD.Low}}{This is the currency pair low price.}
    \item{\code{EURGBP.Close}}{This is the currency pair closing price.}
    \item{\code{EURGBP.High}}{This is the currency pair high price.}
    \item{\code{EURGBP.Low}}{This is the currency pair low price.}
    \item{\code{EURJPY.Close}}{This is the currency pair closing price.}
    \item{\code{EURJPY.High}}{This is the currency pair high price.}
    \item{\code{EURJPY.Low}}{This is the currency pair low price.}
    \item{\code{EURCHF.Close}}{This is the currency pair closing price.}
    \item{\code{EURCHF.High}}{This is the currency pair high price.}
    \item{\code{EURCHF.Low}}{This is the currency pair low price.}
    \item{\code{AUDUSD.Close}}{This is the currency pair closing price.}
    \item{\code{AUDUSD.High}}{This is the currency pair high price.}
    \item{\code{AUDUSD.Low}}{This is the currency pair low price.}
    \item{\code{GBPJPY.Close}}{This is the currency pair closing price.}
    \item{\code{GBPJPY.High}}{This is the currency pair high price.}
    \item{\code{GBPJPY.Low}}{This is the currency pair low price.}
    \item{\code{CHFJPY.Close}}{This is the currency pair closing price.}
    \item{\code{CHFJPY.High}}{This is the currency pair high price.}
    \item{\code{CHFJPY.Low}}{This is the currency pair low price.}
    \item{\code{GBPCHF.Close}}{This is the currency pair closing price.}
    \item{\code{GBPCHF.High}}{This is the currency pair high price.}
    \item{\code{GBPCHF.Low}}{This is the currency pair low price.}
    \item{\code{NZDUSD.Close}}{This is the currency pair closing price.}
    \item{\code{NZDUSD.High}}{This is the currency pair high price.}
    \item{\code{NZDUSD.Low}}{This is the currency pair low price.}
    }
}
\source{\url{https://www.global-view.com/forex-trading-tools/forex-history/index.html}}
\keyword{datasets}
