\name{data.demonsessions}
\alias{demonsessions}
\title{Demon Sessions Data Set}
\usage{data(demonsessions)}
\description{
  These are the monthly number of user sessions at
  \url{https://web.archive.org/web/20141224051720/http://www.bayesian-inference.com/index} by continent. Additional data
  may be added in the future.
}
\format{
  This data frame contains 26 rows (with row names) and 6 columns. The
  following data dictionary describes each variable or column.
  \describe{
    \item{\code{Africa}}{This is the African continent.}
    \item{\code{Americas}}{This is North and South America.}
    \item{\code{Asia}}{This is the Asian continent.}
    \item{\code{Europe}}{This is Europe as a continent.}
    \item{\code{Oceania}}{This is Oceania, such as Australia.}
    \item{\code{Not.Set}}{This includes sessions in which the continent
      was not set, or is unknown.}
    }
}
\source{\url{https://web.archive.org/web/20141224051720/http://www.bayesian-inference.com/index}}
\keyword{datasets}
