\name{data.demonsnacks}
\alias{demonsnacks}
\title{Demon Snacks Data Set}
\usage{data(demonsnacks)}
\description{
  Late one night, after witnessing Laplace's Demon in action, I followed
  him back to what seemed to be his lair. Minutes later, he left again.
  I snuck inside and saw something labeled 'Demon Snacks'. Hurriedly, I
  recorded the 39 items, each with a name and 10 nutritional attributes.
}
\format{
  This data frame contains 39 rows (with row names) and 10 columns. The
  following data dictionary describes each variable or column.
  \describe{
    \item{\code{Serving.Size}}{This is serving size in grams.}
    \item{\code{Calories}}{This is the number of calories.}
    \item{\code{Total.Fat}}{This is total fat in grams.}
    \item{\code{Saturated.Fat}}{This is saturated fat in grams.}
    \item{\code{Cholesterol}}{This is cholesterol in milligrams.}
    \item{\code{Sodium}}{This is sodium in milligrams.}
    \item{\code{Total.Carbohydrate}}{This is the total carbohydrates in grams.}
    \item{\code{Dietary.Fiber}}{This is dietary fiber in grams.}
    \item{\code{Sugars}}{This is sugar in grams.}
    \item{\code{Protein}}{This is protein in grams.}
    }
}
\source{This data was obtained from the lair of Laplace's Demon!}
\keyword{datasets}
