\name{data.demontexas}
\alias{demontexas}
\title{Demon Space-Time Data Set}
\usage{data(demontexas)}
\description{
  This data set is for space-time models that require latitude and
  longitude, or coordinates. This data set consists of the minimum,
  mean, and maximum temperatures in Texas for 13 months.
}
\format{
  This data frame contains 369 rows of sites in Texas and 43
  columns. The following data dictionary describes each variable or
  column.
  \describe{
    \item{\code{Elevation}}{This is the elevation of the site.}
    \item{\code{Latitude}}{This is the latitude of the site.}
    \item{\code{Longitude}}{This is the longitude of the site.}
    \item{\code{Gulf}}{This is a gulf indicator of the site.}
    \item{\code{Max1}}{This is the maximum temperature in month 1.}
    \item{\code{Max2}}{This is the maximum temperature in month 2.}
    \item{\code{Max3}}{This is the maximum temperature in month 3.}
    \item{\code{Max4}}{This is the maximum temperature in month 4.}
    \item{\code{Max5}}{This is the maximum temperature in month 5.}
    \item{\code{Max6}}{This is the maximum temperature in month 6.}
    \item{\code{Max7}}{This is the maximum temperature in month 7.}
    \item{\code{Max8}}{This is the maximum temperature in month 8.}
    \item{\code{Max9}}{This is the maximum temperature in month 9.}
    \item{\code{Max10}}{This is the maximum temperature in month 10.}
    \item{\code{Max11}}{This is the maximum temperature in month 11.}
    \item{\code{Max12}}{This is the maximum temperature in month 12.}
    \item{\code{Max13}}{This is the maximum temperature in month 13.}
    \item{\code{Mean1}}{This is the mean temperature in month 1.}
    \item{\code{Mean2}}{This is the mean temperature in month 2.}
    \item{\code{Mean3}}{This is the mean temperature in month 3.}
    \item{\code{Mean4}}{This is the mean temperature in month 4.}
    \item{\code{Mean5}}{This is the mean temperature in month 5.}
    \item{\code{Mean6}}{This is the mean temperature in month 6.}
    \item{\code{Mean7}}{This is the mean temperature in month 7.}
    \item{\code{Mean8}}{This is the mean temperature in month 8.}
    \item{\code{Mean9}}{This is the mean temperature in month 9.}
    \item{\code{Mean10}}{This is the mean temperature in month 10.}
    \item{\code{Mean11}}{This is the mean temperature in month 11.}
    \item{\code{Mean12}}{This is the mean temperature in month 12.}
    \item{\code{Mean13}}{This is the mean temperature in month 13.}
    \item{\code{Min1}}{This is the minimum temperature in month 1.}
    \item{\code{Min2}}{This is the minimum temperature in month 2.}
    \item{\code{Min3}}{This is the minimum temperature in month 3.}
    \item{\code{Min4}}{This is the minimum temperature in month 4.}
    \item{\code{Min5}}{This is the minimum temperature in month 5.}
    \item{\code{Min6}}{This is the minimum temperature in month 6.}
    \item{\code{Min7}}{This is the minimum temperature in month 7.}
    \item{\code{Min8}}{This is the minimum temperature in month 8.}
    \item{\code{Min9}}{This is the minimum temperature in month 9.}
    \item{\code{Min10}}{This is the minimum temperature in month 10.}
    \item{\code{Min11}}{This is the minimum temperature in month 11.}
    \item{\code{Min12}}{This is the minimum temperature in month 12.}
    \item{\code{Min13}}{This is the minimum temperature in month 13.}
    }
}
\source{\url{http://www.stat.ufl.edu/~winner/datasets.html}}
\keyword{datasets}
