\name{LaplacesDemon-package}
\alias{LaplacesDemon-package}
\docType{package}
\title{The LaplacesDemon Package}
\description{
  Welcome to LaplacesDemon, a complete environment for Bayesian
  inference within R.
}
\details{
  \tabular{ll}{
    Package: \tab LaplacesDemon\cr
    Type: \tab Package\cr
    Version: \tab 14.04.23\cr
    Date: \tab 2014-04-23\cr
    License: \tab MIT\cr
  }
  The goal of LaplacesDemon, often referred to as Laplace's Demon, is to
  provide a complete and self-contained Bayesian environment within R.
  For example, this package includes dozens of MCMC algorithms, Laplace
  Approximation, iterative quadrature, variational Bayes,
  parallelization, big data, PMC, over 90 examples in the ``Examples''
  vignette, dozens of additional probability distributions, numerous
  MCMC diagnostics, Bayes factors, posterior predictive checks, a
  variety of plots, elicitation, parameter and variable importance,
  Bayesian forms of test statistics (such as Durbin-Watson, Jarque-Bera,
  etc.), validation, and numerous additional utility functions, such as
  functions for multimodality, matrices, or timing your model
  specification. Other vignettes include an introduction to Bayesian
  inference, as well as a tutorial.

  There are many plans for the growth of this package, and many are
  long-term plans such as to cotinuously stockpile distributions,
  examples, samplers, and optimization algorithms. Contributions to this
  package are welcome.

  The main function in this package is the \code{LaplacesDemon}
  function, and the best place to start is probably with the
  LaplacesDemon Tutorial vignette.
  }
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\keyword{package}