\name{print.heidelberger}
\alias{print.heidelberger}
\title{Print an object of class \code{heidelberger} to the screen.}
\description{
  This may be used to print the contents of an object of class
  \code{heidelberger} to the screen.
}
\usage{\method{print}{heidelberger}(x, digits=3, \dots)}
\arguments{
  \item{x}{An object of class \code{heidelberger} is required.}
  \item{digits}{This is the number of digits to print.}
  \item{\dots}{Additional arguments are unused.}
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{
  \code{\link{Heidelberger.Diagnostic}}.
}
\examples{### See the Heidelberger.Diagnostic function for an example.}
\keyword{print}
