\name{print.iterquad}
\alias{print.iterquad}
\title{Print an object of class \code{iterquad} to the screen.}
\description{
  This may be used to print the contents of an object of class
  \code{iterquad} to the screen.
}
\usage{\method{print}{iterquad}(x, \dots)}
\arguments{
  \item{x}{An object of class \code{iterquad} is required.}
  \item{\dots}{Additional arguments are unused.}
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{\code{\link{IterativeQuadrature}}}
\examples{### See the IterativeQuadrature function for an example.}
\keyword{print}