\name{print.laplace}
\alias{print.laplace}
\title{Print an object of class \code{laplace} to the screen.}
\description{
  This may be used to print the contents of an object of class
  \code{laplace} to the screen.
}
\usage{\method{print}{laplace}( x, \dots)}
\arguments{
     \item{x}{An object of class \code{laplace} is required.}
     \item{\dots}{Additional arguments are unused.}
}
\author{Byron Hall \email{statisticat@gmail.com}}
\seealso{\code{\link{LaplaceApproximation}}}
