\name{print.miss}
\alias{print.miss}
\title{Print an object of class \code{miss} to the screen.}
\description{
  This may be used to print the contents of an object of class
  \code{miss} to the screen.
}
\usage{\method{print}{miss}(x, \dots)}
\arguments{
  \item{x}{An object of class \code{miss} is required.}
  \item{\dots}{Additional arguments are unused.}
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{
  \code{\link{MISS}}.
}
\examples{### See the MISS function for an example.}
\keyword{print}
