\name{print.pmc}
\alias{print.pmc}
\title{Print an object of class \code{pmc} to the screen.}
\description{
  This may be used to print the contents of an object of class
  \code{pmc} to the screen.
}
\usage{\method{print}{pmc}(x, \dots)}
\arguments{
  \item{x}{An object of class \code{pmc} is required.}
  \item{\dots}{Additional arguments are unused.}
}
\references{
  Hall, B. (2012). "Laplace's Demon", STATISTICAT, LLC.
  URL=\url{http://www.statisticat.com/laplacesdemon.html}
}
\author{Byron Hall \email{statisticat@gmail.com}}
\seealso{
  \code{\link{PMC}}.
}
\examples{### See the PMC function for an example.}
\keyword{print}
