\name{print.raftery}
\alias{print.raftery}
\title{Print an object of class \code{raftery} to the screen.}
\description{
  This may be used to print the contents of an object of class
  \code{raftery} to the screen.
}
\usage{\method{print}{raftery}(x, digits=3, \dots)}
\arguments{
  \item{x}{An object of class \code{raftery} is required.}
  \item{digits}{This is the number of digits to print.}
  \item{\dots}{Additional arguments are unused.}
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{
  \code{\link{Raftery.Diagnostic}}.
}
\examples{### See the Raftery.Diagnostic function for an example.}
\keyword{print}
