\name{print.vb}
\alias{print.vb}
\title{Print an object of class \code{vb} to the screen.}
\description{
  This may be used to print the contents of an object of class
  \code{vb} to the screen.
}
\usage{\method{print}{vb}(x, \dots)}
\arguments{
  \item{x}{An object of class \code{vb} is required.}
  \item{\dots}{Additional arguments are unused.}
}
\author{Statisticat, LLC. \email{software@bayesian-inference.com}}
\seealso{\code{\link{VariationalBayes}}}
\examples{### See the VariationalBayes function for an example.}
\keyword{print}