\name{LaplacesDemon-package}
\alias{LaplacesDemon-package}
\docType{package}
\title{The LaplacesDemon Package}
\description{
Welcome to LaplacesDemon, an R package for Bayesian inference.
}
\details{
  \tabular{ll}{
    Package: \tab LaplacesDemon\cr
    Type: \tab Package\cr
    Version: \tab 11.12.05\cr
    Date: \tab 2011-12-05\cr
    License: \tab GPL-2\cr
  }
  The goal of LaplacesDemon, often referred to as Laplace's Demon, is to
  provide a complete and self-contained Bayesian environment within R.
  For example, this package includes numerous MCMC algorithms, Laplace
  Approximation with multiple optimization algorithms, over 70 examples
  in the ``Examples'' vignette, dozens of additional probability
  distributions, numerous MCMC diagnostics, Bayes factors, posterior
  predictive checks, a variety of plots, elicitation, parameter and
  variable importance, and numerous additional utility functions, such
  as functions for multimodality, matrices, or timing your model
  specification. Other vignettes include an introduction to Bayesian
  inference, as well as a tutorial.

  There are many plans for the growth of this package, and many are
  long-term plans such as to cotinuously stockpile distributions,
  examples, samplers, and optimization algorithms. Contributions to this
  package are welcome.

  The main function in this package is the \code{LaplacesDemon}
  function, and the best place to start is probably with the
  LaplacesDemon Tutorial vignette.
  }
\references{
  Hall, B. (2011). "Laplace's Demon", STATISTICAT, LLC.
  URL=\url{http://www.statisticat.com/laplacesdemon.html}
}
\author{Byron Hall \email{laplacesdemon@statisticat.com}}
\keyword{package}