\name{LaplacesDemon-package}
\alias{LaplacesDemon-package}
\alias{.colVars}
\alias{.iqagh}
\alias{.iqaghsg}
\alias{.iqcagh}
\alias{.laaga}
\alias{.labfgs}
\alias{.labhhh}
\alias{.lacg}
\alias{.ladfp}
\alias{.lahar}
\alias{.lahj}
\alias{.lalbfgs}
\alias{.lalm}
\alias{.lanm}
\alias{.lanr}
\alias{.lapso}
\alias{.larprop}
\alias{.lasgd}
\alias{.lasoma}
\alias{.laspg}
\alias{.lasr1}
\alias{.latr}
\alias{.mcmcadmg}
\alias{.mcmcafss}
\alias{.mcmcagg}
\alias{.mcmcahmc}
\alias{.mcmcaies}
\alias{.mcmcam}
\alias{.mcmcamm}
\alias{.mcmcamm.b}
\alias{.mcmcamwg}
\alias{.mcmccharm}
\alias{.mcmcdemc}
\alias{.mcmcdram}
\alias{.mcmcdrm}
\alias{.mcmcess}
\alias{.mcmcgibbs}
\alias{.mcmcgg}
\alias{.mcmcggcp}
\alias{.mcmcggcpp}
\alias{.mcmcggdp}
\alias{.mcmcggdpp}
\alias{.mcmcharm}
\alias{.mcmchmc}
\alias{.mcmchmcda}
\alias{.mcmcim}
\alias{.mcmcinca}
\alias{.mcmcmala}
\alias{.mcmcmcmcmc}
\alias{.mcmcmtm}
\alias{.mcmcmwg}
\alias{.mcmcnuts}
\alias{.mcmcohss}
\alias{.mcmcram}
\alias{.mcmcrdmh}
\alias{.mcmcrefractive}
\alias{.mcmcrj}
\alias{.mcmcrss}
\alias{.mcmcrwm}
\alias{.mcmcsamwg}
\alias{.mcmcsgld}
\alias{.mcmcslice}
\alias{.mcmcsmwg}
\alias{.mcmcthmc}
\alias{.mcmctwalk}
\alias{.mcmcuess}
\alias{.mcmcusamwg}
\alias{.mcmcusmwg}
\alias{.rowVars}
\alias{.vbsalimans2}
\docType{package}
\title{
\packageTitle{LaplacesDemon}
}
\description{
\packageDescription{LaplacesDemon}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LaplacesDemon}
\packageIndices{LaplacesDemon}

  The goal of LaplacesDemon, often referred to as LD, is to provide a
  complete and self-contained Bayesian environment within R. For
  example, this package includes dozens of MCMC algorithms, Laplace
  Approximation, iterative quadrature, variational Bayes,
  parallelization, big data, PMC, over 100 examples in the ``Examples''
  vignette, dozens of additional probability distributions, numerous
  MCMC diagnostics, Bayes factors, posterior predictive checks, a
  variety of plots, elicitation, parameter and variable importance,
  Bayesian forms of test statistics (such as Durbin-Watson, Jarque-Bera,
  etc.), validation, and numerous additional utility functions, such as
  functions for multimodality, matrices, or timing your model
  specification. Other vignettes include an introduction to Bayesian
  inference, as well as a tutorial.

  No further development of this package is currently being done as the original
  maintainer has stopped working on the package. Contributions to this
  package are welcome at \url{https://github.com/LaplacesDemonR/LaplacesDemon}.

  The main function in this package is the \code{LaplacesDemon}
  function, and the best place to start is probably with the
  LaplacesDemon Tutorial vignette.
  }
\author{
\packageAuthor{LaplacesDemon}

Maintainer: \packageMaintainer{LaplacesDemon}
}
\keyword{package}
