\name{print.demonoid}
\alias{print.demonoid}
\title{Print an object of class \code{demonoid} to the screen.}
\description{
  This may be used to print the contents of an object of class
  \code{demonoid} to the screen.
}
\usage{\method{print}{demonoid}(x, \dots)}
\arguments{
  \item{x}{An object of class \code{demonoid} is required.}
  \item{\dots}{Additional arguments are unused.}
}
\references{
  Hall, B. (2012). "Laplace's Demon", STATISTICAT, LLC.
  URL=\url{http://www.statisticat.com/laplacesdemon.html}
}
\author{Byron Hall \email{laplacesdemon@statisticat.com}}
\seealso{
  \code{\link{Consort}},
  \code{\link{LaplacesDemon}}.
}
\examples{### See the LaplacesDemon function for an example.}
\keyword{print}