\name{print.laplace}
\alias{print.laplace}
\title{Print an object of class \code{laplace} to the screen.}
\description{
  This may be used to print the contents of an object of class
  \code{laplace} to the screen.
}
\usage{\method{print}{laplace}(x, \dots)}
\arguments{
     \item{x}{An object of class \code{laplace} is required.}
     \item{\dots}{Additional arguments are unused.}
}
\references{
  Hall, B. (2011). "Laplace's Demon", STATISTICAT, LLC.
  URL=\url{http://www.statisticat.com/laplacesdemon.html}
}
\author{Byron Hall \email{laplacesdemon@statisticat.com}}
\seealso{\code{\link{LaplaceApproximation}}}
\examples{### See the LaplaceApproximation function for an example.}
\keyword{print}